[Maths] refactor TScalarVector<>
Previous impl had some serious issues: could read trashed memory with
the previous expr unfolding, and general meta programming hell to handle
type promotion with hair pulling problems around overloading confusions
and other hacks that made intellisense want to shot itself...

Something has to be made, so:
- components are now trivial scalar, eg. bare float, to avoid confusion
with parameter type deduction
- TScalarVectorExpr<> is explicitly referencing the wrapped type, still
to simply promotion and general overloading
- every component/swizzle has a valid representation of the data
- still using TScalarVectorExpr<> with CRTP
- still support swizzling
- still support mixing scalars and vectors in operators/function calls
- still support expr unfolding, but always expand the result to a vector
if the expression requires promotion

+++ Source/Runtime/Core/Private/Allocator/MallocBinned2.cpp
@@ -650,8 +650,8 @@ FBinnedSmallTable::FBinnedSmallTable() NOEXCEPT {
         pool.BlockSize = checked_cast<u16>(FMallocBinned2::SmallPoolIndexToBlockSize(i));
         pool.MaxChunksEverAllocated = 0;
         pool.MaxBlocksPerBundle = checked_cast<u8>(Min( // bundle size doesn't depend on chunk size
-            PPE_MALLOCBINNED2_BUNDLE_MAX_COUNT,
-            (PPE_MALLOCBINNED2_BUNDLE_MAX_SIZE + pool.BlockSize - 1u) / pool.BlockSize ));
+            static_cast<size_t>(PPE_MALLOCBINNED2_BUNDLE_MAX_COUNT),
+            static_cast<size_t>((PPE_MALLOCBINNED2_BUNDLE_MAX_SIZE + pool.BlockSize - 1u) / pool.BlockSize) ));
         Assert_NoAssume(pool.MaxBlocksPerBundle > 0);
 
         u32 bestWastedSize = UMax, bestPagesPerChunk = 0;
