+++ Source/Runtime/RHI/Public/RHI/BufferDesc.h
@@ -12,18 +12,18 @@ namespace RHI {
 //////////////////////////////////////////////////////////////////////////////
 //----------------------------------------------------------------------------
 struct FBufferDesc {
-    u32 SizeInBytes{ 0 };
+    size_t SizeInBytes{ 0 };
     EBufferUsage Usage{ Default };
     EQueueUsage Queues{ Default };
 
     FBufferDesc() = default;
-    FBufferDesc(u32 sizeInBytes, EBufferUsage usage, EQueueUsage queues = Default)
+    FBufferDesc(size_t sizeInBytes, EBufferUsage usage, EQueueUsage queues = Default)
     :   SizeInBytes(sizeInBytes)
     ,   Usage(usage)
     ,   Queues(queues)
     {}
 
-    FBufferDesc& SetSize(u32 sz) { SizeInBytes = sz; return (*this); }
+    FBufferDesc& SetSize(size_t sz) { SizeInBytes = sz; return (*this); }
     FBufferDesc& SetUsage(EBufferUsage usage) { Usage = usage; return (*this); }
     FBufferDesc& SetQueues(EQueueUsage queues) { Queues = queues; return (*this); }
 
@@ -32,11 +32,11 @@ PPE_ASSUME_TYPE_AS_POD(FBufferDesc)
 //----------------------------------------------------------------------------
 struct FBufferViewDesc {
     EPixelFormat Format{ Default };
-    u32 Offset{ UMax };
-    u32 SizeInBytes{ UMax };
+    size_t Offset{ UMax };
+    size_t SizeInBytes{ UMax };
 
     FBufferViewDesc() = default;
-    FBufferViewDesc(EPixelFormat format, u32 offset, u32 sizeInBytes)
+    FBufferViewDesc(EPixelFormat format, size_t offset, size_t sizeInBytes)
     :   Format(format)
     ,   Offset(offset)
     ,   SizeInBytes(sizeInBytes)
